+++ source/patterns/@aws-solutions-constructs/core/lib/kinesis-streams-helper.ts
@@ -16,6 +16,7 @@ import * as kinesis from '@aws-cdk/aws-kinesis';
 import { DefaultStreamProps } from './kinesis-streams-defaults';
 import * as cdk from '@aws-cdk/core';
 import { overrideProps } from './utils';
+import * as cloudwatch from '@aws-cdk/aws-cloudwatch';
 
 export interface BuildKinesisStreamProps {
   /**
@@ -30,12 +31,10 @@ export interface BuildKinesisStreamProps {
    *
    * @default - Default props are used.
    */
-  readonly kinesisStreamProps?: kinesis.StreamProps | any
+  readonly kinesisStreamProps?: kinesis.StreamProps
 }
 
-export function buildKinesisStream(scope: cdk.Construct, props?: BuildKinesisStreamProps): kinesis.Stream {
-    // If props is undefined, define it
-    props = (props === undefined) ? {} : props;
+export function buildKinesisStream(scope: cdk.Construct, props: BuildKinesisStreamProps): kinesis.Stream {
 
     if (props.existingStreamObj) {
         return props.existingStreamObj;
@@ -53,4 +52,39 @@ export function buildKinesisStream(scope: cdk.Construct, props?: BuildKinesisStr
 
     // Create the stream and return
     return new kinesis.Stream(scope, 'KinesisStream', kinesisStreamProps);
+}
+
+export function buildKinesisStreamCWAlarms(scope: cdk.Construct): cloudwatch.Alarm[] {
+    // Setup CW Alarms for KinesisStream
+    const alarms: cloudwatch.Alarm[] = new Array();
+
+    // Alarm if Max (GetRecords.IteratorAgeMilliseconds): >= 12 hours (50% of 24 hours default retention period)
+    alarms.push(new cloudwatch.Alarm(scope, 'KinesisStreamGetRecordsIteratorAgeAlarm', {
+      metric: new cloudwatch.Metric({
+        namespace: 'AWS/Kinesis',
+        metricName: 'GetRecords.IteratorAgeMilliseconds'
+      }),
+      threshold: 2592000, // 12 Hours (50% of 24 hours - default record retention period)
+      evaluationPeriods: 1,
+      statistic: 'Maximum',
+      period: cdk.Duration.minutes(5),
+      comparisonOperator: cloudwatch.ComparisonOperator.GREATER_THAN_OR_EQUAL_TO_THRESHOLD,
+      alarmDescription: 'Consumer Record Processing Falling Behind, there is risk for data loss due to record expiration.'
+    }));
+
+    // Alarm if Avg (ReadProvisionedThroughputExceeded): > 0
+    alarms.push(new cloudwatch.Alarm(scope, 'KinesisStreamReadProvisionedThroughputExceededAlarm', {
+        metric: new cloudwatch.Metric({
+          namespace: 'AWS/Kinesis',
+          metricName: 'ReadProvisionedThroughputExceeded'
+        }),
+        threshold: 0,
+        evaluationPeriods: 1,
+        statistic: 'Average',
+        period: cdk.Duration.minutes(5),
+        comparisonOperator: cloudwatch.ComparisonOperator.GREATER_THAN_THRESHOLD,
+        alarmDescription: 'Consumer Application is Reading at a Slower Rate Than Expected.'
+    }));
+
+    return alarms;
 }