+++ source/patterns/@aws-solutions-constructs/core/test/lambda-event-source.test.ts
@@ -12,30 +12,56 @@
  */
 
 import * as defaults from '../index';
-import { DynamoEventSourceProps, KinesisEventSourceProps } from '@aws-cdk/aws-lambda-event-sources';
+import { DynamoEventSourceProps, KinesisEventSourceProps, SqsDlq } from '@aws-cdk/aws-lambda-event-sources';
 import * as lambda from '@aws-cdk/aws-lambda';
 import * as s3 from '@aws-cdk/aws-s3';
 import '@aws-cdk/assert/jest';
+import { Duration, Stack } from '@aws-cdk/core';
 
-test('test DynamoEventSourceProps', () => {
-  const props = defaults.DynamoEventSourceProps();
+test('test DynamoEventSourceProps with defaults', () => {
+  const stack = new Stack();
+  const props = defaults.DynamoEventSourceProps(stack);
+
+  expect(props.onFailure).toBeInstanceOf(SqsDlq);
+  expect(props.startingPosition).toEqual("TRIM_HORIZON");
+  expect(props.bisectBatchOnError).toEqual(true);
+  expect(props.retryAttempts).toEqual(500);
+  expect(props.maxRecordAge).toEqual(Duration.hours(24));
+});
+
+test('test DynamoEventSourceProps with deployDeadLetterQueue: false', () => {
+  const stack = new Stack();
+  const props = defaults.DynamoEventSourceProps(stack, {
+    deploySqsDlqQueue: false
+  });
 
   expect(props).toEqual({
-    startingPosition: "TRIM_HORIZON"
+    startingPosition: "TRIM_HORIZON",
+    bisectBatchOnError: true,
+    maxRecordAge: Duration.hours(24),
+    retryAttempts: 500
   });
 });
 
 test('test DynamoEventSourceProps override', () => {
+  const stack = new Stack();
   const myProps: DynamoEventSourceProps = {
     startingPosition: lambda.StartingPosition.LATEST,
-    batchSize: 1
+    batchSize: 5,
+    retryAttempts: 3
   };
 
-  const props = defaults.DynamoEventSourceProps(myProps);
+  const props = defaults.DynamoEventSourceProps(stack, {
+    eventSourceProps: myProps,
+    deploySqsDlqQueue: false
+  });
 
   expect(props).toEqual({
-    batchSize: 1,
-    startingPosition: "LATEST"
+    batchSize: 5,
+    startingPosition: "LATEST",
+    maxRecordAge: Duration.hours(24),
+    bisectBatchOnError: true,
+    retryAttempts: 3
   });
 });
 
@@ -58,28 +84,65 @@ test('test S3EventSourceProps w/ user props', () => {
   });
 });
 
-test('test KinesisEventSourceProps', () => {
-  const props = defaults.KinesisEventSourceProps();
+test('test KinesisEventSourceProps with defaults', () => {
+  const stack = new Stack();
+  const props = defaults.KinesisEventSourceProps(stack);
+
+  expect(props.onFailure).toBeInstanceOf(SqsDlq);
+  expect(props.startingPosition).toEqual("TRIM_HORIZON");
+  expect(props.bisectBatchOnError).toEqual(true);
+  expect(props.retryAttempts).toEqual(500);
+  expect(props.maxRecordAge).toEqual(Duration.hours(24));
+});
+
+test('test KinesisEventSourceProps with deployDeadLetterQueue: false', () => {
+  const stack = new Stack();
+  const props = defaults.KinesisEventSourceProps(stack, {
+    deploySqsDlqQueue: false
+  });
 
   expect(props).toEqual({
     startingPosition: "TRIM_HORIZON",
-    bisectBatchOnError: true
+    bisectBatchOnError: true,
+    maxRecordAge: Duration.hours(24),
+    retryAttempts: 500
   });
 });
 
 test('test KinesisEventSourceProps override', () => {
+  const stack = new Stack();
   const myProps: KinesisEventSourceProps = {
     startingPosition: lambda.StartingPosition.LATEST,
     batchSize: 5,
     retryAttempts: 3
   };
 
-  const props = defaults.KinesisEventSourceProps(myProps);
+  const props = defaults.KinesisEventSourceProps(stack, {
+    eventSourceProps: myProps,
+    deploySqsDlqQueue: false
+  });
 
   expect(props).toEqual({
     batchSize: 5,
     startingPosition: "LATEST",
+    maxRecordAge: Duration.hours(24),
     bisectBatchOnError: true,
     retryAttempts: 3
+  });
+});
+
+test('test sqsDlqQueueProps override', () => {
+  const stack = new Stack();
+
+  defaults.KinesisEventSourceProps(stack, {
+    sqsDlqQueueProps: {
+      queueName: 'hello-world',
+      visibilityTimeout: Duration.seconds(50)
+    }
+  });
+
+  expect(stack).toHaveResource("AWS::SQS::Queue", {
+      QueueName: "hello-world",
+      VisibilityTimeout: 50
   });
 });